requirements.txt
web3
pwntools
pycryptodome
requests
PyJWT
numpy